Moved test compression providers & compressors to a separate class
